Fix issue where game object packets would crash client.
This issue was caused by the instance incorrectly calculating the
player's surrounding entities (of all types, we got lucky detecting this
bug early on).

Region sizes were decreased to 16x16 (as opposed to 32x32) as the
server would not send the object position packet on time. This caused
some areas to not display objects at all. I have also increased the
default radius of game and wall objects that the player knows at any
given time to 35x35. This seems to fit well in testing.

Minor refactoring was done in this commit as well, such as changing the
name of 'wall decoration' to 'wall object.' The ObjectUpdatePosition has
a new function, acknowledge(), which acknowledges that the player knows
of a new/removed object in their view. The intersection() function was
also slightly modified to be more readable.

Lastly, the game object position sender now has 'packet overflow'
detection in it. This is needed, for whatever reason, in case the
packet's size grows higher than 5000 bytes. 5000 bytes is the maximum
size of a packet that the client can read at once. This mechanism needs
to be implemented for ALL entity update packets (and, in general,
packets that have potential to overflow).