Implemented packet buffering, refactored handlers
Incoming packets are now buffered. If a message arrives with X length,
but only Y bytes are avaiable to read (assuming X > Y), then the message
contents are stashed. The next time the client sends any data, it is
appended to the stashed buffer until X bytes have been read.

Once X bytes has been read, the payload of the packet is then read,
along with its ID. Finally, a `packet` event is emmited to the session
where it can be handled.

Packet handlers are now async functions rather than Promises (which was
redundant to begin with).